+++ src/wai/annotations/imgvis/isp/annotation_overlay/specifier/__init__.py
@@ -1,1 +1,2 @@
+from ._AnnotationOverlayICISPSpecifier import AnnotationOverlayICISPSpecifier
 from ._AnnotationOverlayODISPSpecifier import AnnotationOverlayODISPSpecifier