+++ src/wai/annotations/imgvis/sink/image_viewer/specifier/__init__.py
@@ -1,1 +1,3 @@
-from ._ImageViewerSinkSpecifier import ImageViewerSinkSpecifier
+from ._ImageViewerICSinkSpecifier import ImageViewerICSinkSpecifier
+from ._ImageViewerISSinkSpecifier import ImageViewerISSinkSpecifier
+from ._ImageViewerODSinkSpecifier import ImageViewerODSinkSpecifier