+++ CHANGES.rst
@@ -4,5 +4,5 @@ Changelog
 1.0.0 (????-??-??)
 ------------------
 
-- Initial release: add-annotation-overlay-od, image-viewer
+- Initial release: add-annotation-overlay-(ic/od), image-viewer-(ic/is/od)